+++ src/Pim/Bundle/CatalogBundle/Manager/ProductManager.php
@@ -3,6 +3,10 @@
 namespace Pim\Bundle\CatalogBundle\Manager;
 
 use Doctrine\Common\Persistence\ObjectManager;
+use Doctrine\Common\Util\ClassUtils;
+use Pim\Component\Resource\Model\SaverInterface;
+use Pim\Component\Resource\Model\BulkSaverInterface;
+use Pim\Component\Resource\Model\RemoverInterface;
 use Pim\Bundle\CatalogBundle\Builder\ProductBuilder;
 use Pim\Bundle\CatalogBundle\Entity\Repository\AssociationTypeRepository;
 use Pim\Bundle\CatalogBundle\Entity\Repository\AttributeOptionRepository;
@@ -27,7 +31,7 @@ use Symfony\Component\EventDispatcher\GenericEvent;
  * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ProductManager
+class ProductManager implements SaverInterface, BulkSaverInterface, RemoverInterface
 {
     /** @var array */
     protected $configuration;
@@ -176,14 +180,19 @@ class ProductManager
      *
      * @param ProductInterface    $product
      * @param AvailableAttributes $availableAttributes
-     *
-     * @return null
+     * @param array               $savingOptions
      */
-    public function addAttributesToProduct(ProductInterface $product, AvailableAttributes $availableAttributes)
-    {
+    public function addAttributesToProduct(
+        ProductInterface $product,
+        AvailableAttributes $availableAttributes,
+        array $savingOptions = []
+    ) {
         foreach ($availableAttributes->getAttributes() as $attribute) {
             $this->builder->addAttributeToProduct($product, $attribute);
         }
+
+        $options = array_merge(['recalculate' => false, 'schedule' => false], $savingOptions);
+        $this->save($product, $options);
     }
 
     /**
@@ -191,47 +200,36 @@ class ProductManager
      *
      * @param ProductInterface  $product
      * @param AbstractAttribute $attribute
-     *
-     * @return boolean
+     * @param array             $savingOptions
      */
-    public function removeAttributeFromProduct(ProductInterface $product, AbstractAttribute $attribute)
-    {
-        $this->builder->removeAttributeFromProduct($product, $attribute);
+    public function removeAttributeFromProduct(
+        ProductInterface $product,
+        AbstractAttribute $attribute,
+        array $savingOptions = []
+    ) {
+        foreach ($product->getValues() as $value) {
+            if ($attribute === $value->getAttribute()) {
+                $product->removeValue($value);
+            }
+        }
+
+        $options = array_merge(['recalculate' => false, 'schedule' => false], $savingOptions);
+        $this->save($product, $options);
     }
 
     /**
-     * Save a product
-     *
-     * @param ProductInterface $product     The product to save
-     * @param boolean          $recalculate Whether or not to directly recalculate the completeness
-     * @param boolean          $flush       Whether or not to flush the entity manager
-     * @param boolean          $schedule    Whether or not to schedule the product for completeness recalculation
-     *
-     * @return null
-     *
-     * @deprecated use saveProduct() instead. Will be removed in 1.3
+     * {@inheritdoc}
      */
-    public function save(ProductInterface $product, $recalculate = true, $flush = true, $schedule = true)
+    public function save($object, array $options = [])
     {
-        $options = [
-            'recalculate' => $recalculate,
-            'flush' => $flush,
-            'schedule' => $schedule,
-        ];
-
-        return $this->saveProduct($product, $options);
-    }
-
-    /**
-     * Save a product
-     *
-     * @param ProductInterface $product The product to save
-     * @param array            $options Saving options
-     *
-     * @return null
-     */
-    public function saveProduct(ProductInterface $product, array $options = [])
-    {
+        if (!$object instanceof ProductInterface) {
+            throw new \InvalidArgumentException(
+                sprintf(
+                    'Expects a Pim\Bundle\CatalogBundle\Model\ProductInterface, "%s" provided',
+                    ClassUtils::getClass($object)
+                )
+            );
+        }
         $options = array_merge(
             [
                 'recalculate' => true,
@@ -241,38 +239,13 @@ class ProductManager
             $options
         );
 
-        return $this->persister->persist($product, $options);
+        $this->persister->persist($object, $options);
     }
 
     /**
-     * Save multiple products
-     *
-     * @param ProductInterface[] $products    The products to save
-     * @param boolean            $recalculate Whether or not to directly recalculate the completeness
-     * @param boolean            $flush       Whether or not to flush the entity manager
-     * @param boolean            $schedule    Whether or not to schedule the product for completeness recalculation
-     *
-     * @return null
-     * @deprecated use saveAllProducts() instead. Will be removed in 1.3
+     * {@inheritdoc}
      */
-    public function saveAll(array $products, $recalculate = false, $flush = true, $schedule = true)
-    {
-        $options = [
-            'recalculate' => $recalculate,
-            'flush' => $flush,
-            'schedule' => $schedule,
-        ];
-
-        return $this->saveAllProducts($products, $options);
-    }
-
-    /**
-     * Save multiple products
-     *
-     * @param ProductInterface[] $products The products to save
-     * @param array              $options  Saving options
-     */
-    public function saveAllProducts(array $products, array $options = [])
+    public function saveAll(array $objects, array $options = [])
     {
         $allOptions = array_merge(
             [
@@ -285,13 +258,39 @@ class ProductManager
         $itemOptions = $allOptions;
         $itemOptions['flush'] = false;
 
-        foreach ($products as $product) {
-            $this->saveProduct($product, $itemOptions);
+        foreach ($objects as $object) {
+            $this->save($object, $itemOptions);
         }
 
-        if ($allOptions['flush'] === true) {
+        if (true === $allOptions['flush']) {
             $this->objectManager->flush();
         }
+    }
+
+    /**
+     * Save a product
+     *
+     * @param ProductInterface $product The product to save
+     * @param array            $options Saving options
+     *
+     * @deprecated will be removed in 1.4, use save()
+     */
+    public function saveProduct(ProductInterface $product, array $options = [])
+    {
+        $this->save($product, $options);
+    }
+
+    /**
+     * Save multiple products
+     *
+     * @param ProductInterface[] $products The products to save
+     * @param array              $options  Saving options
+     *
+     * @deprecated will be removed in 1.4, use saveAll()
+     */
+    public function saveAllProducts(array $products, array $options = [])
+    {
+        $this->saveAll($products, $options);
     }
 
     /**
@@ -434,32 +433,25 @@ class ProductManager
     }
 
     /**
-     * Remove products
-     *
-     * @param integer[] $ids
+     * {@inheritdoc}
      */
-    public function removeAll(array $ids)
+    public function remove($object, array $options = [])
     {
-        $products = $this->getProductRepository()->findByIds($ids);
-        foreach ($products as $product) {
-            $this->remove($product, false);
+        if (!$object instanceof ProductInterface) {
+            throw new \InvalidArgumentException(
+                sprintf(
+                    'Expects a Pim\Bundle\CatalogBundle\Model\ProductInterface, "%s" provided',
+                    ClassUtils::getClass($object)
+                )
+            );
         }
-        $this->objectManager->flush();
-    }
 
-    /**
-     * Remove a product
-     *
-     * @param ProductInterface $product
-     * @param boolean          $flush
-     */
-    public function remove(ProductInterface $product, $flush = true)
-    {
-        $this->eventDispatcher->dispatch(ProductEvents::PRE_REMOVE, new GenericEvent($product));
-        $this->objectManager->remove($product);
-        $this->eventDispatcher->dispatch(ProductEvents::POST_REMOVE, new GenericEvent($product));
+        $options = array_merge(['flush' => true], $options);
+        $this->eventDispatcher->dispatch(ProductEvents::PRE_REMOVE, new GenericEvent($object));
+        $this->objectManager->remove($object);
+        $this->eventDispatcher->dispatch(ProductEvents::POST_REMOVE, new GenericEvent($object));
 
-        if (true === $flush) {
+        if (true === $options['flush']) {
             $this->objectManager->flush();
         }
     }