+++ src/Pim/Bundle/EnrichBundle/Controller/AssociationTypeController.php
@@ -8,7 +8,7 @@ use Pim\Bundle\CatalogBundle\Entity\AssociationType;
 use Pim\Bundle\CatalogBundle\Manager\AssociationManager;
 use Pim\Bundle\CatalogBundle\Manager\AssociationTypeManager;
 use Pim\Bundle\EnrichBundle\AbstractController\AbstractDoctrineController;
-use Pim\Bundle\EnrichBundle\Form\Handler\AssociationTypeHandler;
+use Pim\Bundle\EnrichBundle\Form\Handler\HandlerInterface;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
 use Symfony\Bundle\FrameworkBundle\Templating\EngineInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
@@ -30,7 +30,7 @@ use Symfony\Component\Validator\ValidatorInterface;
  */
 class AssociationTypeController extends AbstractDoctrineController
 {
-    /** @var AssociationTypeHandler */
+    /** @var HandlerInterface */
     protected $assocTypeHandler;
 
     /** @var Form */
@@ -56,7 +56,7 @@ class AssociationTypeController extends AbstractDoctrineController
      * @param ManagerRegistry          $doctrine
      * @param AssociationTypeManager   $assocTypeManager
      * @param AssociationManager       $assocManager
-     * @param AssociationTypeHandler   $assocTypeHandler
+     * @param HandlerInterface         $assocTypeHandler
      * @param Form                     $assocTypeForm
      */
     public function __construct(
@@ -71,7 +71,7 @@ class AssociationTypeController extends AbstractDoctrineController
         ManagerRegistry $doctrine,
         AssociationTypeManager $assocTypeManager,
         AssociationManager $assocManager,
-        AssociationTypeHandler $assocTypeHandler,
+        HandlerInterface $assocTypeHandler,
         Form $assocTypeForm
     ) {
         parent::__construct(