+++ src/Pim/Bundle/EnrichBundle/Manager/SequentialEditManager.php
@@ -3,6 +3,9 @@
 namespace Pim\Bundle\EnrichBundle\Manager;
 
 use Doctrine\Common\Persistence\ObjectManager;
+use Doctrine\Common\Util\ClassUtils;
+use Pim\Component\Resource\Model\SaverInterface;
+use Pim\Component\Resource\Model\RemoverInterface;
 use Pim\Bundle\CatalogBundle\Manager\ProductManager;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Pim\Bundle\EnrichBundle\Entity\Repository\SequentialEditRepository;
@@ -17,7 +20,7 @@ use Symfony\Component\Security\Core\User\UserInterface;
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class SequentialEditManager
+class SequentialEditManager implements SaverInterface, RemoverInterface
 {
     /** @var ObjectManager */
     protected $om;
@@ -52,14 +55,21 @@ class SequentialEditManager
     }
 
     /**
-     * Save a sequential edit entity
-     *
-     * @param SequentialEdit $sequentialEdit
+     * {@inheritdoc}
      */
-    public function save(SequentialEdit $sequentialEdit)
+    public function save($object, array $options = [])
     {
-        $this->om->persist($sequentialEdit);
-        $this->om->flush($sequentialEdit);
+        if (!$object instanceof SequentialEdit) {
+            throw new \InvalidArgumentException(
+                sprintf(
+                    'Expects a Pim\Bundle\EnrichBundle\Entity\SequentialEdit, "%s" provided',
+                    ClassUtils::getClass($object)
+                )
+            );
+        }
+
+        $this->om->persist($object);
+        $this->om->flush($object);
     }
 
     /**
@@ -76,14 +86,21 @@ class SequentialEditManager
     }
 
     /**
-     * Remove a sequential edit entity
-     *
-     * @param SequentialEdit $sequentialEdit
+     * {@inheritdoc}
      */
-    public function remove(SequentialEdit $sequentialEdit)
+    public function remove($object, array $options = [])
     {
-        $this->om->remove($sequentialEdit);
-        $this->om->flush($sequentialEdit);
+        if (!$object instanceof SequentialEdit) {
+            throw new \InvalidArgumentException(
+                sprintf(
+                    'Expects a Pim\Bundle\EnrichBundle\Entity\SequentialEdit, "%s" provided',
+                    ClassUtils::getClass($object)
+                )
+            );
+        }
+
+        $this->om->remove($object);
+        $this->om->flush($object);
     }
 
     /**